requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py[voice]
redis
psycopg2-binary